refactor: marshal && unmarshal

+++ internal/community/instance.go
@@ -31,7 +31,7 @@ func BindAccount(hashedAccount string, publicKey *string) error {
 
 	rpcAddress := conf.GetNode().ExternalAddr
 	rpcPort := conf.GetNode().ExternalPort
-	version := 0
+	version := uint32(0)
 
 	if err := storage.UpsertMember(hashedAccount, *publicKey, privateKeyValut, rpcAddress, rpcPort, &version); err != nil {
 		return err
